pyproject.toml
[project]
name = "moondream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.0",
    "pillow>=11.0.0",
    "timm>=1.0.11",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.46.2",
]
